[#3] Resolve "Import AWS SDK"
# New feature and improvement

This MR introduces the AWS SDK for writing and reading inside an S3 bucket.
In order to use it, you need to import the code from the `aws-integration` module and update the dependencies of your specific provisioner's main module.
You need also to specify an `is-mock` parameter inside `reference.conf` that is telling you if you want a mocked version of the aws SDK or not. In case of not, you need to set up some Environment variables in order to authenticate to AWS:

- `AWS_REGION`: the region in which the S3 bucket is created
- `AWS_ACCESS_KEY_ID`: taken from your aws account
- `AWS_SECRET_ACCESS_KEY`: taken from your aws account
- `AWS_SESSION_TOKEN`: taken from your aws account (only in case of temporary authentication)

# Related Issue

Closes #3